pipeline/requirements.txt
# N'Ko Training Pipeline Requirements

# HTTP & Async
aiohttp>=3.9.0
httpx>=0.25.0

# Google Gemini API
google-generativeai>=0.3.0

# Web scraping
beautifulsoup4>=4.12.0
scrapetube>=2.5.0

# Data handling
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Audio processing (optional)
pydub>=0.25.0

# Supabase client
supabase>=2.0.0
postgrest>=0.13.0

# Utilities
tqdm>=4.66.0
tenacity>=8.2.0
